convert TokenStream using Source.queue
